apindex: complete rewrite
Pretty much everything except the template HTMLs were rewritten.

Now, it requires the output of `tree -Js` to function. This can either
be a good thing or bad thing; Windows may not work anymore. On the
other hand, this allows flexibility in that HTML files need not be in
the actual directory of the files to function; this may be useful if
one uses an identical directory structure, but the index files are
served on another host (i.e. assets on S3, index somewhere else).

Features added:
- Allow a different base URL
    - Useful if actual files are hosted on a different host than the
      index.html files
- Allow outputting the HTMLs to a directory of a user's choice
    - By default, it will use the `$PWD/site` directory
    - One can even pick the directory that the `tree -Js` command was
      run, if the files and the indexes are co-hosted

Features removed:
- Last modified time is removed
    - Purely due to laziness on my part; `tree` doesn't seem to have a
      structured output for modified time, the behaviour being
      inconsistent depending on the mod time
    - Support for it may be added at some point in the future