Directory: add the `curpath` argument
The intention was simply to replace the path that each file links to,
not to display the full backend path at the top as well.

This design  decision is because, if a site is hosted on `example.com`,
but the base URL is set to `contoso.com`, it will likely confuse the
end user who intended to visit `example.com` only to see a totally
different base URL displayed to them.

With this change, it allows to have a different file path to display to
the end user while retaining the redirect to another host for the real
file.