Absolute path-ify all hrefs
This is a breaking change: all paths are now absolute paths. This is
needed for some weird production environments.

To do this, use the previously-added `curpath` variable and implement
it in File class.

Make sure we don't prepend the path with `.` when dealing with this.

A CLI argument has been added for users who wish to publish the
generated files into a subdirectory.