Fix baseurl behaviour
baseurl should only be added to the script if specified. the previous behaviour resulted in files having links such as /ace3ds.com/old/ace3ds.com/old/ace3ds.com_Original_Kernel_20120308.zip for example. we also need to add .count(True) to the end of the ignoredextension condition as otherwise, no matter what the resulting values are, it's treated as though the condition is true (???)